requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.26.0
pytest>=7.0.0
streamlit>=1.33.0
plotly>=5.18.0
pandas>=2.0.0
google-analytics-data>=0.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
